feat(auth): don't delete an unverified account with an active subscription
Because:

* We currently delete unverified accounts in a number of places, but with [FXA-3486](https://mozilla-hub.atlassian.net/browse/FXA-3486), it will be possible for an unverified account to have an active subscription.

This commit:

* Ensures we don't delete unverified accounts that have active subscriptions.

Closes #8745

+++ packages/fxa-auth-server/lib/email/notifications.js
@@ -5,6 +5,8 @@
 'use strict';
 
 const utils = require('./utils/helpers');
+const { default: Container } = require('typedi');
+const { StripeHelper } = require('../../lib/payments/stripe');
 
 // Account deletion threshold for new unverified accounts that receive
 // a bounce or complaint notification. Unverified accounts younger than
@@ -12,6 +14,7 @@ const utils = require('./utils/helpers');
 const SIX_HOURS = 1000 * 60 * 60 * 6;
 
 module.exports = (log, error) => {
+  const stripeHelper = Container.get(StripeHelper);
   return (queue, db) => {
     queue.start();
 
@@ -45,7 +48,8 @@ module.exports = (log, error) => {
 
               if (
                 !emailRecord.emailVerified &&
-                emailRecord.createdAt >= Date.now() - SIX_HOURS
+                emailRecord.createdAt >= Date.now() - SIX_HOURS &&
+                !(await stripeHelper.hasActiveSubscription(emailRecord.uid))
               ) {
                 // A bounce or complaint on a new unverified account is grounds for deletion
                 await db.deleteAccount(emailRecord);
